+++ react/components/EmptyState.tsx
@@ -8,10 +8,17 @@ const CSS_HANDLES = ['noPickupsStateContent'] as const
 
 interface Props {
   onClick?: () => void
-  title: string
-  description: string
-  buttonLabel: string
+  title?: string
+  description?: string
+  buttonLabel?: string
   variant?: 'primary' | 'secondary'
+  iconProps?: IconProps
+}
+
+interface IconProps {
+  width?: string
+  height?: string
+  useIcon?: boolean
 }
 
 const EmptyState = ({
@@ -20,6 +27,7 @@ const EmptyState = ({
   description,
   buttonLabel,
   variant = 'primary',
+  iconProps = { useIcon: true },
 }: Props) => {
   const handle = useCssHandles(CSS_HANDLES)
 
@@ -27,13 +35,21 @@ const EmptyState = ({
     <div className="flex-auto flex flex-column">
       <div
         style={{ margin: 'auto 0' }}
-        className={`flex flex-column ${
+        className={`flex ${title ? 'flex-column' : 'flex-row'} ${
           variant === 'primary' ? 'items-center' : 'items-start'
         } ${handle.noPickupsStateContent}`}
       >
-        <PinSlashIcon color={variant === 'primary' ? '#979899' : '#000000'} />
-        <p className="f3 fw6 mb0 mt5">{title}</p>
-        <p className="t-small mt3 mb0 c-muted-1">{description}</p>
+        {iconProps.useIcon && (
+          <PinSlashIcon
+            color={variant === 'primary' ? '#979899' : '#000000'}
+            width={iconProps?.width}
+            height={iconProps?.height}
+          />
+        )}
+        {title && <p className="f3 fw6 mb0 mt5">{title}</p>}
+        {description && (
+          <p className="t-small mt1 ml3 mb0 c-muted-1">{description}</p>
+        )}
       </div>
       {onClick && (
         <div className="mt8">