+++ react/typings/gql.d.ts
@@ -1,5 +1,5 @@
 declare module '*.gql' {
-  import { DocumentNode } from 'graphql'
+  import type { DocumentNode } from 'graphql'
 
   const Schema: DocumentNode
 
@@ -7,7 +7,7 @@ declare module '*.gql' {
 }
 
 declare module '*.graphql' {
-  import { DocumentNode } from 'graphql'
+  import type { DocumentNode } from 'graphql'
 
   const schema: DocumentNode