+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `LocationDetectorButton` component to automatically detect
+
 ## [1.8.0] - 2025-10-06
 
 ### Added